# API-Docs.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:9090",
	"basePath": "/",
	"tags": [
		{
			"name": "contact-rest-controller",
			"description": "Contact Rest Controller"
		}
	],
	"paths": {
		"/contact": {
			"post": {
				"tags": [
					"contact-rest-controller"
				],
				"summary": "saveContact",
				"operationId": "saveContactUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "contactid",
						"in": "query",
						"required": false,
						"type": "integer",
						"format": "int32"
					},
					{
						"name": "email",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "name",
						"in": "query",
						"required": false,
						"type": "string"
					},
					{
						"name": "phoneno",
						"in": "query",
						"required": false,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/contact/{id}": {
			"get": {
				"tags": [
					"contact-rest-controller"
				],
				"summary": "getContact",
				"operationId": "getContactUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Contact"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			},
			"delete": {
				"tags": [
					"contact-rest-controller"
				],
				"summary": "deleteContact",
				"operationId": "deleteContactUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			}
		},
		"/contacts": {
			"get": {
				"tags": [
					"contact-rest-controller"
				],
				"summary": "ViewAllContacts",
				"operationId": "ViewAllContactsUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Contact"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		}
	},
	"definitions": {
		"Contact": {
			"type": "object",
			"properties": {
				"contactid": {
					"type": "integer",
					"format": "int32"
				},
				"email": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"phoneno": {
					"type": "integer",
					"format": "int64"
				}
			},
			"title": "Contact"
		}
	}
}